pyproject.toml
# SPDX-FileCopyrightText: 2022 Contributors to the Power Grid Model IO project <dynamic.grid.calculation@alliander.com>
#
# SPDX-License-Identifier: MPL-2.0

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "power-grid-model-io"
authors = [{name = "Alliander Dynamic Grid Calculation", email = "dynamic.grid.calculation@alliander.com"}]
description = "Power Grid Model Input/Output"
readme = "README.md"
keywords = ["power grid model", "input/output", "conversions"]
license = {text = "MPL-2.0"}
classifiers=[
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Topic :: Scientific/Engineering :: Physics",
]
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "pandas",
    "power_grid_model>=1.4rc",
    "structlog",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "pylint",
    "pytest",
    "pytest-cov",
]
tabular = [
    "pyyaml",
]
excel = [
    "openpyxl",
]
cli = [
    "typer[all]",
]
all = [
    "openpyxl",
    "pyyaml",
    "typer[all]",
]

[project.urls]
Home-page = "https://github.com/alliander-opensource/power-grid-model-io"

[project.scripts]
convert = "power_grid_model_io.cli.convert:app"
validate = "power_grid_model_io.cli.validate:app"

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[tool.setuptools.dynamic]
version = {file = "PYPI_VERSION"}

[tool.pytest.ini_options]
testpaths = ["tests/unit"]
addopts = ["--cov=src/power_grid_model_io", "--cov-report", "term", "--cov-fail-under=20"]

[tool.black]
line-length = 120
target-version = ['py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 120

[tool.pylint]
max-line-length = 120

[tool.mypy]
follow_imports = "silent"
ignore_missing_imports = true
show_column_numbers = true
non_interactive = true
install_types = true
